- Cleanup various warnings     - WINSCR: Implement new configuration dialog with new parameters     - WINSCR: Remove crufty code     - WINSCR: Update copyrights in resource files
    client/
        app.cpp
    clientscr/
        boinc_ss.h
        boinc_ss.rc
        boinc_ss_opengl.h
        boinc_ss_opengl.rc
    clientscr/
        screensaver.cpp, .h
        screensaver_win.cpp, .h
    lib/
        common_defs.h
        crypt.cpp
        idlemon_win.cpp
        win_util.cpp
    samples/worker/
        worker.cpp
    win_build/
        boinc_ss.vcproj
        multi_thread.vcproj
        sleeper.vcproj
        ss_app.vcproj
        uc2_graphics.vcproj

svn path=/trunk/boinc/; revision=19923

+++ client/app.cpp
@@ -810,7 +810,7 @@ void MSG_QUEUE::msg_queue_poll(MSG_CHANNEL& channel) {
 // if the last message in the buffer is "msg", remove it and return 1
 //
 int MSG_QUEUE::msg_queue_purge(const char* msg) {
-	int count = msgs.size();
+	int count = (int)msgs.size();
 	if (!count) return 0;
 	vector<string>::iterator iter = msgs.begin();
 	for (int i=0; i<count-1; i++) {
